Refactor ingest path: Split sql_ingest.go into mutiple files.
