+++ pkg/pgmodel/ingestor.go
@@ -1,56 +1,40 @@
-// This file and its contents are licensed under the Apache License 2.0.
-// Please see the included NOTICE for copyright information and
-// LICENSE for a copy of the license.
-
 package pgmodel
 
 import (
 	"fmt"
-	"strconv"
 
+	"github.com/timescale/promscale/pkg/clockcache"
+	"github.com/timescale/promscale/pkg/pgxconn"
 	"github.com/timescale/promscale/pkg/prompb"
 )
 
-const (
-	MetricNameLabelName = "__name__"
-)
-
-var (
-	ErrNoMetricName = fmt.Errorf("metric name missing")
-)
-
-// SeriesID represents a globally unique id for the series. This should be equivalent
-// to the PostgreSQL type in the series table (currently BIGINT).
-type SeriesID int64
-
-func (s SeriesID) String() string {
-	return strconv.FormatInt(int64(s), 10)
-}
-
-// inserter is responsible for inserting label, series and data into the storage.
-type inserter interface {
-	InsertNewData(rows map[string][]samplesInfo) (uint64, error)
-	CompleteMetricCreation() error
-	Close()
-}
-
-// SeriesCache provides a caching mechanism for labels and series.
-type SeriesCache interface {
-	GetSeries(lset Labels) (SeriesID, error)
-	SetSeries(lset Labels, id SeriesID) error
-	NumElements() int
-	Capacity() int
-}
-
-type samplesInfo struct {
-	labels   *Labels
-	seriesID SeriesID
-	samples  []prompb.Sample
+type Cfg struct {
+	AsyncAcks       bool
+	ReportInterval  int
+	SeriesCacheSize uint64
+	NumCopiers      int
 }
 
 // DBIngestor ingest the TimeSeries data into Timescale database.
 type DBIngestor struct {
 	db inserter
+}
+
+// NewPgxIngestorWithMetricCache returns a new Ingestor that uses connection pool and a metrics cache
+// for caching metric table names.
+func NewPgxIngestorWithMetricCache(conn pgxconn.PgxConn, cache MetricCache, cfg *Cfg) (*DBIngestor, error) {
+	pi, err := newPgxInserter(conn, cache, cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &DBIngestor{db: pi}, nil
+}
+
+// NewPgxIngestor returns a new Ingestor that write to PostgreSQL using PGX
+func NewPgxIngestor(conn pgxconn.PgxConn) (*DBIngestor, error) {
+	cache := &MetricNameCache{clockcache.WithMax(DefaultMetricCacheSize)}
+	return NewPgxIngestorWithMetricCache(conn, cache, &Cfg{})
 }
 
 // Ingest transforms and ingests the timeseries data into Timescale database.
@@ -67,7 +51,7 @@ func (i *DBIngestor) Ingest(tts []prompb.TimeSeries, req *prompb.WriteRequest) (
 
 	rowsInserted, err := i.db.InsertNewData(data)
 	if err == nil && int(rowsInserted) != totalRows {
-		return rowsInserted, fmt.Errorf("Failed to insert all the data! Expected: %d, Got: %d", totalRows, rowsInserted)
+		return rowsInserted, fmt.Errorf("failed to insert all the data! Expected: %d, Got: %d", totalRows, rowsInserted)
 	}
 	return rowsInserted, err
 }
